backend/requirements.txt
# Python 3.11 Backend Dependencies for Interview Vault
# FastAPI Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Environment & Configuration
python-dotenv==1.0.0

# Email validation (required by Pydantic EmailStr)
email-validator>=2.0.0

# HTTP Client (for API calls) - version must be compatible with supabase/gotrue
# gotrue uses proxy parameter which requires httpx>=0.27.0
httpx>=0.27.0

# Supabase Python Client
supabase>=2.0.0,<3.0.0

# DNS Validation
dnspython==2.5.0

# Template Engine (for email HTML)
Jinja2==3.1.3

# Form Data Parsing
python-multipart==0.0.6

# APScheduler for background jobs (optional, for local cron)
apscheduler==3.10.4

# AI Providers
openai>=1.0.0
